Remove containers because they add too much noise
The concept of containers is sound, but having two files for every component is too much boilerplate. The you enter a directory, the component that shares the name of that directory is the container. When that file starts appearing too large, it should mean it's time to break it up in to smaller pieces.